WT-18509 Move the stable btree ID lookup into the disaggregated test helper
Two layered schema tests each carried their own copy. Keep one in the
mixin alongside the stable configuration lookup it builds on.
